Add files via upload
This version is the final version of the software and fixes an issue with unclear naming of values in the output. The name of each value in the output is now described in detail. Fixed the problem that the software may report errors under windows system. And the code is packaged by pyinstaller to run on windows and macOS respectively.